实现遗传算法寻路

+++ Assets/Scripts/GNN/AI/NeuralNetworks/MathHelper.cs
@@ -5,21 +5,21 @@ using UnityEngine;
 
 public class MathHelper{
 
-    public static double SigmoidFunction(double xValue)
+    public static float SigmoidFunction(float xValue)
     {
-        if (xValue > 10) return 1.0;
-        else if (xValue < -10) return 0.0;
-        else return 1.0 / (1.0 + Math.Exp(-xValue));
+        if (xValue > 10) return 1.0f;
+        else if (xValue < -10) return 0.0f;
+        else return (float)(1.0 / (1.0 + Math.Exp(-xValue)));
     }
 
-    public static double TanHFunction(double xValue)
+    public static float TanHFunction(float xValue)
     {
-        if (xValue > 10) return 1.0;
-        else if (xValue < -10) return -1.0;
-        else return Math.Tanh(xValue);
+        if (xValue > 10) return 1.0f;
+        else if (xValue < -10) return -1.0f;
+        else return (float)Math.Tanh(xValue);
     }
 
-    public static double SoftSignFunction(double xValue)
+    public static float SoftSignFunction(float xValue)
     {
         return xValue / (1 + Math.Abs(xValue));
     }
